Run the code via meta + r
Intercept the keyboard refresh command when the code is in focus and use it to send an ajax call to run the code. The initial hieght of the code window has been increased to 20em, this should make it slightly easier when dealing with large chunks of code. I've also fixed a bug with the line numbers not correctly updating when the user clicks on the code window and gains focus.